+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 -----
 
 * BREAKING: the ``lod`` key is required in the YAML configuration file
+* creates the ``cjdb_multipolygon_to_multisurface(geometry)`` SQL function
 
 0.1.0 (2019-12-16)
 ------------------